+++ service/controller/cert.go
@@ -5,12 +5,13 @@ import (
 	"fmt"
 	"time"
 
-	corev1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/core/v1alpha1"
-	providerv1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/provider/v1alpha1"
-	"github.com/giantswarm/k8sclient"
+	corev1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/core/v1alpha1"
+	providerv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/provider/v1alpha1"
+	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	"github.com/giantswarm/operatorkit/controller"
+	"github.com/giantswarm/operatorkit/v4/pkg/controller"
+	"github.com/giantswarm/operatorkit/v4/pkg/resource"
 	"github.com/giantswarm/vaultcrt"
 	"github.com/giantswarm/vaultpki"
 	"github.com/giantswarm/vaultpki/key"
@@ -19,12 +20,14 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	apiv1alpha2 "sigs.k8s.io/cluster-api/api/v1alpha2"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/cert-operator/pkg/label"
+	"github.com/giantswarm/cert-operator/pkg/project"
 )
 
 type CertConfig struct {
@@ -95,7 +98,7 @@ func NewCert(config CertConfig) (*Cert, error) {
 		}
 	}
 
-	var ResourceSet *controller.ResourceSet
+	var resources []resource.Interface
 	{
 		c := ResourceSetConfig{
 			K8sClient:   config.K8sClient.K8sClient(),
@@ -110,7 +113,7 @@ func NewCert(config CertConfig) (*Cert, error) {
 			ProjectName:         config.ProjectName,
 		}
 
-		ResourceSet, err = NewResourceSet(c)
+		resources, err = NewResourceSet(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -122,9 +125,10 @@ func NewCert(config CertConfig) (*Cert, error) {
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
 			Name:      config.ProjectName,
-			ResourceSets: []*controller.ResourceSet{
-				ResourceSet,
-			},
+			Resources: resources,
+			Selector: labels.SelectorFromSet(map[string]string{
+				label.OperatorVersion: project.Version(),
+			}),
 			NewRuntimeObjectFunc: func() runtime.Object {
 				return new(corev1alpha1.CertConfig)
 			},